start work on prioritization

+++ src/Event.java
@@ -2,7 +2,7 @@ import java.time.*;
 import java.util.*;
 
 public abstract class Event {
-	protected int priority;
+	protected double priority;
 	protected LocalDateTime dateTimeCreated;
 	protected String name;
 	protected ArrayList<LocalDateTime> breakdown;
